refactor: change structure folders
